% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCCC.R
\name{gg_ccc}
\alias{gg_ccc}
\title{Plot of raw data with line of best fit and CCC}
\usage{
gg_ccc(data, x, y, hor = "left", ver = "top")
}
\arguments{
\item{data}{dataframe.}

\item{x, y}{numeric vectors or characters. If numeric vectors, then they specify the x- and y-
plotting variables, respectively. If characters, then they are taken to specify
the columns containing the x- and y-variables in \code{data}. If not provided,
then the x- and y-plotting variables are taken as the first two columsns of \code{data}.}

\item{hor}{'left' or 'right'. Specifies horizontal position of summary statistic table.
Default is \code{'left'}.}

\item{ver}{'top' or 'bottom'. Specifies vertical position of summary statistic table.
Default is \code{'top'}.}
}
\value{
A \code{ggplot2} plot with the following elements:
- raw data plotted
- y ~ x linear line of best fit (estimate + confidence bands)
- table of relevant summary statistics.
}
\description{
Plot of raw data with line of best fit and CCC
}
\details{
The following summary statistics are
printed on the plot:
- Concordance CC: est (95% CI)
- Pearon's CC: est (95% CI)
- Intercept: est (95% CI)
- Slope: est (95% CI)
}
