% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcorr.R
\name{ggcorr}
\alias{ggcorr}
\title{Plot scatterplot with CCC and ab-line overlaid}
\usage{
ggcorr(data, title, trans = NULL)
}
\arguments{
\item{data}{dataframe. Has columns group and y, which
specify the grouping variable and response variable, respectively.}

\item{trans}{'sqrt' or 'log'. Transformation to be applied before
both calculating CCC and displaying data.}
}
\description{
Plot scatterplot with CCC and ab-line overlaid

character. Title, appended by the CCC estimate and 95% CI.
}
