% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_ccc.R
\name{gg_ccc}
\alias{gg_ccc}
\title{Plot of raw data with line of best fit and CCC}
\usage{
gg_ccc(data, x, y, hor = "left", ver = "top", shift_x = 0,
  shift_y = 0, table_font_size = 5, table_font_gap = 2.5)
}
\arguments{
\item{data}{dataframe.}

\item{x, y}{numeric vectors or characters. If numeric vectors, then they specify the x- and y-
plotting variables, respectively. If characters, then they are taken to specify
the columns containing the x- and y-variables in \code{data}. If not provided,
then the x- and y-plotting variables are taken as the first two columsns of \code{data}.
\code{x} and \code{y} must have the same type, and anything else generates an error.}

\item{hor}{'left' or 'right'. Specifies horizontal position of summary statistic table.
Default is \code{'left'}.}

\item{ver}{'top' or 'bottom'. Specifies vertical position of summary statistic table.
Default is \code{'top'}.}

\item{shift_x}{numeric. Percentage of x-axis range to shift summary statistic table.
If positive, then the table is shifted to the right. For example, \code{shift_x=50}
means that the table will be shifted halfway across the width of the x-axis range.
Default is zero.}

\item{shift_y}{numeric. Analogous to \code{shift_x}, but for the y-axis. Positive
values move upwards. Default is 0.}

\item{table_font_size}{numeric. Size of font for summary statistic table.}

\item{table_font_gap}{numeric. Percentage of range of y-axis that rows in table are
separated by.}
}
\value{
A \code{ggplot2} plot with the following elements:
\itemize{
  \item raw data plotted
  \item black 45 degree line
  \item blue linear line of best fit (estimate + confidence bands)
  \item table of relevant summary statistics.
}
}
\description{
Plot of raw data with line of best fit and CCC
}
\details{
The following summary statistics are
printed on the plot:
\itemize{
  \item Concordance CC: est (95\% CI)
  \item Pearson's CC: est (95\% CI)
  \item Intercept: est (95\% CI)
  \item Slope: est (95\% CI)
}
}
\examples{
data( test_tbl )
gg_ccc( test_tbl )
}
